refactor(web-ui): extract shared styled-text renderer
