+++ web-ui/src/components/mod.rs
@@ -6,5 +6,6 @@ pub mod login;
 pub mod nick_list;
 pub mod shell_view;
 pub mod status_line;
+pub mod styled;
 pub mod topic_bar;
 pub mod wizard;
